server should be able to connect to an alien mom
