+++ src/include/pbs_nodes.h
@@ -484,7 +484,8 @@ extern  void	momptr_offline_by_mom(mominfo_t *, char *);
 extern  void	momptr_clear_offline_by_mom(mominfo_t *, char *);
 extern  void	   delete_mom_entry(mominfo_t *);
 extern  mominfo_t *create_svrmom_entry(char *, unsigned int, unsigned long *);
-extern  void       delete_svrmom_entry(mominfo_t *);
+extern  void	delete_svrmom_entry(mominfo_t *);
+extern  int	create_svrmom_struct(pbs_node *pnode);
 extern  int	legal_vnode_char(char, int);
 extern 	char	*parse_node_token(char *, int, int *, char *);
 extern  int	cross_link_mom_vnode(struct pbsnode *, mominfo_t *);
